+++ gems/aws-sdk-s3/spec/client_spec.rb
@@ -20,12 +20,18 @@ module Aws
         S3::BUCKET_REGIONS.clear
       end
 
-      it 'raises an appropriate error when credentials are missing' do
+      it 'raises an error when credentials are missing' do
         creds = Credentials.new(nil, nil)
         client = Client.new(credentials: creds)
         expect do
           client.list_buckets
         end.to raise_error(Aws::Errors::MissingCredentialsError)
+      end
+
+      it 'raises an error when region is missing' do
+        expect do
+          Client.new(region: nil)
+        end.to raise_error(Aws::Errors::MissingRegionError)
       end
 
       it 'supports empty stub with token set to nil' do